src/requirements.txt
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
requests==2.29.0
six==1.16.0
typing==3.7.4.3
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.3
